Refactor: Cleaning up adjoint codebase
