--- src/gradients/adjoints/discrete_adjoint_gradients.jl
@@ -1,13 +1,0 @@
-module DiscreteAdjoints
-
-using ..OptimalBath: AdjointApproachGradient
-
-include("discrete_adjoint_swe.jl")
-
-export DiscreteAdjointGradient
-
-function DiscreteAdjointGradient(primal_solver::VolumeFluxesSolver)
-    return AdjointApproachGradient(DiscreteAdjointSWE(primal_solver))
-end
-
-end
