+++ src/gradients/adjoints/discrete_adjoint_swe.jl
@@ -1,14 +1,15 @@
-export DiscreteAdjointSWE, TestDiscreteAdjoint, StepWiseTestDiscreteAdjoint
+module DiscreteAdjoints
+
+export DiscreteAdjointSWE
 
 using ForwardDiff: jacobian, gradient
 using VolumeFluxes: CentralUpwind, ShallowWaterEquations1D, XDIR
 using StaticArrays: @SMatrix, SMatrix, setindex
 using ElasticArrays: ElasticArray
 
-import OptimalBath: solve_adjoint
-using .OptimalBath: compute_ghost_cell, AdjointSWE, time_frame
-using .OptimalBath: Grid, for_each_left_boundary_directional_stencil, for_each_interior_directional_stencil, for_each_right_boundary_directional_stencil
-using .OptimalBath: directions
+import OptimalBath
+using OptimalBath
+using OptimalBath: time_frame, compute_ghost_cell
 
 
 struct DiscreteAdjointSWE{PrimalSolver<:VolumeFluxesSolver, GridT, AdjointStates} <: AdjointSWE
@@ -259,7 +260,7 @@ function time_steps(U, ::Grid{Dims}) where Dims
     return size(U, Dims+1)
 end
 
-function solve_adjoint(Λ_end, U::AverageDepthStates, objectives::Objectives, b, t, da::DiscreteAdjointSWE)
+function OptimalBath.solve_adjoint(Λ_end, U::AverageDepthStates, objectives::Objectives, b, t, da::DiscreteAdjointSWE)
     Λ = resize_adjoint_states!(da, U.U)
     return solve_adjoint!(Λ, Λ_end, U, objectives, b, t, da)
 end
@@ -291,4 +292,6 @@ end
         adjoint_pre_proc_step!(Λ, U.U, n-1, grid, da) # pre processing of the next step
     end
     return Λ
+end
+
 end
